// src/app/core/api/model/transferRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TransferRequest { 
    sourceAccountId: string;
    destinationAccountId: string;
    amount: number;
    merchantName?: string;
    description?: string;
    transactionDate: string;
}